Refactor: Babel - Split JSX css prop transform
