+++ packages/babel/src/jsxCssProp/preprocess.ts
@@ -1,0 +1,505 @@
+import { types as t } from "@babel/core";
+import type { NodePath } from "@babel/core";
+import { unwrapTransparentCssRuleExpression } from "../staticCssEval/candidates.js";
+import {
+  findUnsupportedImportedStaticCssEvalReferenceDiagnostic,
+  resolveImportedStaticCssEvalExpression
+} from "../staticCssEval/importedModules.js";
+import { resolveSameFileStaticCssEvalExpression } from "../staticCssEval/sameFile.js";
+import type { PluginState, ProgramScope } from "../types.js";
+import {
+  classifyCssPropValue,
+  containsArraySpreadElement,
+  isArrayClassValueExpression
+} from "./classification.js";
+import { createClassNameAttributeValue } from "./classNameLowering.js";
+import {
+  classNameAttributeName,
+  cssAttributeName,
+  cssExpressionValueErrorMessage,
+  cssModuleHelperImportCleanupScopes,
+  cssModuleName,
+  cssValueErrorMessage,
+  duplicateClassNameErrorMessage,
+  duplicateCssErrorMessage,
+  fragmentTargetErrorMessage,
+  keyRefSpreadErrorMessage,
+  namespacedTargetErrorMessage,
+  unsupportedArraySpreadCssValueErrorMessage,
+  unsupportedDynamicCssRuleValueErrorMessage,
+  unsupportedFunctionCssValueErrorMessage,
+  unsupportedTargetErrorMessage
+} from "./constants.js";
+import {
+  registerImportedStaticCssEvalProviderResultMetadata,
+  registerStaticCssEvalDiagnosticMetadata,
+  registerStaticCssEvalResultMetadata
+} from "./metadata.js";
+import {
+  normalizeResolvedCssRuleExpression,
+  preserveDirectBooleanReferenceValues,
+  resolveDirectInlineStaticCssRuleLiteral,
+  shouldPreserveUnsupportedArraySpreadClassification,
+  shouldPreserveUnsupportedDynamicCssRuleClassification
+} from "./staticCssEval.js";
+import {
+  assertSupportedSpreadAggregationContext,
+  transformSpreadAggregatedCssProp
+} from "./spreadAggregation.js";
+import type { NormalizedJsxCssPropElement } from "./types.js";
+
+export function preprocessJsxCssProp(
+  path: NodePath<t.Program>,
+  state: PluginState
+): boolean {
+  if (state.opts.jsxCssProp !== true) {
+    return false;
+  }
+
+  let transformed = false;
+
+  path.traverse({
+    JSXOpeningElement(openingElementPath) {
+      const normalizedElement = normalizeOpeningElement(
+        openingElementPath,
+        path,
+        state
+      );
+
+      if (!normalizedElement) {
+        return;
+      }
+
+      const { cssAttribute, classNameAttribute } = normalizedElement;
+      if (
+        normalizedElement.hasSpreadBeforeCss ||
+        normalizedElement.hasSpreadAfterCss
+      ) {
+        transformSpreadAggregatedCssProp(openingElementPath, normalizedElement);
+        transformed = true;
+        return;
+      }
+
+      const classNameValue = createClassNameAttributeValue(
+        openingElementPath,
+        normalizedElement
+      );
+
+      const attributes = openingElementPath.node.attributes;
+
+      if (classNameAttribute) {
+        classNameAttribute.value = classNameValue;
+      } else {
+        const cssAttributeIndex = attributes.indexOf(cssAttribute);
+        attributes.splice(
+          cssAttributeIndex,
+          0,
+          t.jsxAttribute(
+            t.jsxIdentifier(classNameAttributeName),
+            classNameValue
+          )
+        );
+      }
+
+      attributes.splice(attributes.indexOf(cssAttribute), 1);
+      transformed = true;
+    }
+  });
+
+  if (transformed) {
+    path.scope.crawl();
+  }
+
+  return transformed;
+}
+
+export function removeUnusedJsxCssPropCssModuleImports(
+  path: NodePath<t.Program>
+): void {
+  const programScope = path.scope as ProgramScope;
+
+  if (!cssModuleHelperImportCleanupScopes.has(programScope)) {
+    return;
+  }
+
+  path.scope.crawl();
+
+  for (const statementPath of path.get("body")) {
+    if (
+      !statementPath.isImportDeclaration() ||
+      statementPath.node.source.value !== cssModuleName
+    ) {
+      continue;
+    }
+
+    const retainedSpecifiers = statementPath.node.specifiers.filter(
+      (specifier) =>
+        !isUnusedCssModuleHelperImportSpecifier(statementPath, specifier)
+    );
+
+    if (retainedSpecifiers.length === statementPath.node.specifiers.length) {
+      continue;
+    }
+
+    if (retainedSpecifiers.length === 0) {
+      statementPath.remove();
+      continue;
+    }
+
+    const nextImportDeclaration = t.cloneNode(statementPath.node);
+    nextImportDeclaration.specifiers = retainedSpecifiers.map((specifier) =>
+      t.cloneNode(specifier)
+    );
+    statementPath.replaceWith(nextImportDeclaration);
+  }
+}
+
+function isUnusedCssModuleHelperImportSpecifier(
+  path: NodePath<t.ImportDeclaration>,
+  specifier: t.ImportDeclaration["specifiers"][number]
+): boolean {
+  if (
+    !t.isImportSpecifier(specifier) ||
+    !t.isIdentifier(specifier.imported) ||
+    (specifier.imported.name !== "css" && specifier.imported.name !== "cx")
+  ) {
+    return false;
+  }
+
+  const binding = path.scope.getBinding(specifier.local.name);
+  return !binding || binding.referencePaths.length === 0;
+}
+
+function normalizeOpeningElement(
+  openingElementPath: NodePath<t.JSXOpeningElement>,
+  programPath: NodePath<t.Program>,
+  state: PluginState
+): NormalizedJsxCssPropElement | null {
+  const { node: openingElement } = openingElementPath;
+  const { attributes } = openingElement;
+  const cssAttributes = getJsxAttributes(openingElement, cssAttributeName);
+
+  if (cssAttributes.length === 0) {
+    return null;
+  }
+
+  assertSupportedCssPropElement(openingElementPath);
+
+  if (cssAttributes.length > 1) {
+    throw openingElementPath.buildCodeFrameError(duplicateCssErrorMessage);
+  }
+
+  const cssAttribute = cssAttributes[0];
+  const cssAttributeIndex = attributes.indexOf(cssAttribute);
+  const attributesBeforeCss = attributes.slice(0, cssAttributeIndex);
+  const attributesAfterCss = attributes.slice(cssAttributeIndex + 1);
+  const hasSpreadBeforeCss = attributesBeforeCss.some((attribute) =>
+    t.isJSXSpreadAttribute(attribute)
+  );
+  const hasSpreadAfterCss = attributesAfterCss.some((attribute) =>
+    t.isJSXSpreadAttribute(attribute)
+  );
+
+  const requiresSpreadAggregation = hasSpreadBeforeCss || hasSpreadAfterCss;
+
+  if (requiresSpreadAggregation && hasExplicitKeyOrRefAttribute(attributes)) {
+    throw openingElementPath.buildCodeFrameError(keyRefSpreadErrorMessage);
+  }
+
+  if (requiresSpreadAggregation) {
+    assertSupportedSpreadAggregationContext(openingElementPath);
+  }
+
+  const classNameAttributes = getJsxAttributes(
+    openingElement,
+    classNameAttributeName
+  );
+
+  if (classNameAttributes.length > 1) {
+    throw openingElementPath.buildCodeFrameError(
+      duplicateClassNameErrorMessage
+    );
+  }
+
+  const cssExpression = getResolvedCssExpression(
+    openingElementPath,
+    programPath,
+    state,
+    cssAttribute
+  );
+  const cssValueClassification = classifyCssPropValue(
+    cssExpression,
+    openingElementPath.scope
+  );
+
+  if (cssValueClassification === "unsupported-function") {
+    throw openingElementPath.buildCodeFrameError(
+      unsupportedFunctionCssValueErrorMessage
+    );
+  }
+
+  if (cssValueClassification === "unsupported-dynamic-css-rule") {
+    throw openingElementPath.buildCodeFrameError(
+      unsupportedDynamicCssRuleValueErrorMessage
+    );
+  }
+
+  if (cssValueClassification === "unsupported-array-spread") {
+    throw openingElementPath.buildCodeFrameError(
+      unsupportedArraySpreadCssValueErrorMessage
+    );
+  }
+
+  return {
+    cssAttribute,
+    cssExpression,
+    cssValueClassification,
+    classNameAttribute: classNameAttributes[0] ?? null,
+    attributesBeforeCss,
+    attributesAfterCss,
+    hasSpreadBeforeCss,
+    hasSpreadAfterCss
+  };
+}
+
+function getResolvedCssExpression(
+  path: NodePath<t.JSXOpeningElement>,
+  programPath: NodePath<t.Program>,
+  state: PluginState,
+  attribute: t.JSXAttribute
+): t.Expression {
+  const cssExpression = getCssExpression(path, attribute);
+  const ownerFile = getStaticCssEvalOwnerFile(state);
+  const staticCssEvalResult = shouldResolveSameFileCssExpression(cssExpression)
+    ? resolveSameFileStaticCssEvalExpression({
+        expression: cssExpression,
+        ownerFile,
+        programPath,
+        scope: path.scope
+      })
+    : { kind: "not-candidate" as const };
+
+  if (staticCssEvalResult.kind === "error") {
+    const inlineStaticCssEvalResult = resolveDirectInlineStaticCssRuleLiteral({
+      expression: cssExpression,
+      ownerFile,
+      programPath,
+      scope: path.scope,
+      provider: state.opts.staticCssEvalProvider,
+      metadata: [],
+      state: { count: 0 },
+      depth: 1
+    });
+
+    if (inlineStaticCssEvalResult.kind === "resolved") {
+      for (const metadata of inlineStaticCssEvalResult.metadata) {
+        registerStaticCssEvalResultMetadata(state, metadata);
+      }
+
+      return preserveDirectBooleanReferenceValues(
+        cssExpression,
+        inlineStaticCssEvalResult.expression
+      );
+    }
+
+    if (inlineStaticCssEvalResult.kind === "error") {
+      if (inlineStaticCssEvalResult.diagnostic.code === "limit-exceeded") {
+        for (const metadata of inlineStaticCssEvalResult.metadata) {
+          registerStaticCssEvalResultMetadata(state, metadata);
+        }
+
+        throw path.buildCodeFrameError(
+          inlineStaticCssEvalResult.diagnostic.message
+        );
+      }
+
+      if (shouldPreserveUnsupportedArraySpreadClassification(cssExpression)) {
+        return cssExpression;
+      }
+
+      if (
+        shouldPreserveUnsupportedDynamicCssRuleClassification(cssExpression)
+      ) {
+        return cssExpression;
+      }
+
+      for (const metadata of inlineStaticCssEvalResult.metadata) {
+        registerStaticCssEvalResultMetadata(state, metadata);
+      }
+
+      throw path.buildCodeFrameError(
+        inlineStaticCssEvalResult.diagnostic.message
+      );
+    }
+
+    if (shouldPreserveUnsupportedArraySpreadClassification(cssExpression)) {
+      return cssExpression;
+    }
+
+    if (shouldPreserveUnsupportedDynamicCssRuleClassification(cssExpression)) {
+      return cssExpression;
+    }
+
+    registerStaticCssEvalResultMetadata(state, staticCssEvalResult);
+    throw path.buildCodeFrameError(staticCssEvalResult.diagnostic.message);
+  }
+
+  if (staticCssEvalResult.kind === "resolved") {
+    registerStaticCssEvalResultMetadata(state, staticCssEvalResult);
+    return preserveDirectBooleanReferenceValues(
+      cssExpression,
+      normalizeResolvedCssRuleExpression(staticCssEvalResult.expression)
+    );
+  }
+
+  registerStaticCssEvalResultMetadata(state, staticCssEvalResult);
+
+  registerImportedStaticCssEvalProviderResultMetadata({
+    expression: cssExpression,
+    ownerFile,
+    state
+  });
+
+  const importedStaticCssEvalResult = resolveImportedStaticCssEvalExpression({
+    expression: cssExpression,
+    ownerFile,
+    provider: state.opts.staticCssEvalProvider,
+    allowUnsupportedSourceFallback: true
+  });
+
+  if (importedStaticCssEvalResult.kind === "error") {
+    registerStaticCssEvalDiagnosticMetadata(
+      state,
+      importedStaticCssEvalResult.diagnostic
+    );
+    throw path.buildCodeFrameError(
+      importedStaticCssEvalResult.diagnostic.message
+    );
+  }
+
+  const resolvedCssExpression =
+    importedStaticCssEvalResult.kind === "resolved"
+      ? importedStaticCssEvalResult.expression
+      : cssExpression;
+
+  const unsupportedImportedReferenceDiagnostic =
+    findUnsupportedImportedStaticCssEvalReferenceDiagnostic({
+      expression: resolvedCssExpression,
+      ownerFile,
+      provider: state.opts.staticCssEvalProvider
+    });
+
+  if (unsupportedImportedReferenceDiagnostic) {
+    registerStaticCssEvalDiagnosticMetadata(
+      state,
+      unsupportedImportedReferenceDiagnostic
+    );
+    throw path.buildCodeFrameError(
+      unsupportedImportedReferenceDiagnostic.message
+    );
+  }
+
+  return resolvedCssExpression;
+}
+
+function getStaticCssEvalOwnerFile(state: PluginState): string {
+  return state.file.opts.filename ?? "<unknown>";
+}
+
+function shouldResolveSameFileCssExpression(expression: t.Expression): boolean {
+  const unwrappedExpression = unwrapTransparentCssRuleExpression(expression);
+
+  return !(
+    t.isArrayExpression(unwrappedExpression) &&
+    !containsArraySpreadElement(unwrappedExpression) &&
+    isArrayClassValueExpression(unwrappedExpression)
+  );
+}
+
+function assertSupportedCssPropElement(
+  openingElementPath: NodePath<t.JSXOpeningElement>
+): void {
+  const { name } = openingElementPath.node;
+
+  if (t.isJSXNamespacedName(name)) {
+    throw openingElementPath.buildCodeFrameError(namespacedTargetErrorMessage);
+  }
+
+  if (isFragmentTarget(name)) {
+    throw openingElementPath.buildCodeFrameError(fragmentTargetErrorMessage);
+  }
+
+  if (t.isJSXIdentifier(name) || t.isJSXMemberExpression(name)) {
+    return;
+  }
+
+  throw openingElementPath.buildCodeFrameError(unsupportedTargetErrorMessage);
+}
+
+function isFragmentTarget(name: t.JSXOpeningElement["name"]): boolean {
+  return (
+    (t.isJSXIdentifier(name) && name.name === "Fragment") ||
+    (t.isJSXMemberExpression(name) &&
+      t.isJSXIdentifier(name.object) &&
+      name.object.name === "React" &&
+      name.property.name === "Fragment")
+  );
+}
+
+function getJsxAttributes(
+  openingElement: t.JSXOpeningElement,
+  attributeName: string
+): t.JSXAttribute[] {
+  return openingElement.attributes.filter(
+    (attribute): attribute is t.JSXAttribute =>
+      t.isJSXAttribute(attribute) &&
+      t.isJSXIdentifier(attribute.name) &&
+      attribute.name.name === attributeName
+  );
+}
+
+function isNamedJsxAttribute(
+  attribute: t.JSXAttribute | t.JSXSpreadAttribute,
+  attributeName: string
+): boolean {
+  return (
+    t.isJSXAttribute(attribute) &&
+    t.isJSXIdentifier(attribute.name) &&
+    attribute.name.name === attributeName
+  );
+}
+
+function hasExplicitKeyOrRefAttribute(
+  attributes: Array<t.JSXAttribute | t.JSXSpreadAttribute>
+): boolean {
+  return attributes.some(
+    (attribute) =>
+      isNamedJsxAttribute(attribute, "key") ||
+      isNamedJsxAttribute(attribute, "ref")
+  );
+}
+
+function getCssExpression(
+  path: NodePath<t.JSXOpeningElement>,
+  attribute: t.JSXAttribute
+): t.Expression {
+  if (attribute.value === null) {
+    throw path.buildCodeFrameError(cssExpressionValueErrorMessage);
+  }
+
+  if (t.isStringLiteral(attribute.value)) {
+    return attribute.value;
+  }
+
+  if (!t.isJSXExpressionContainer(attribute.value)) {
+    throw path.buildCodeFrameError(cssValueErrorMessage);
+  }
+
+  const { expression } = attribute.value;
+
+  if (t.isJSXEmptyExpression(expression)) {
+    throw path.buildCodeFrameError(cssExpressionValueErrorMessage);
+  }
+
+  return expression;
+}
